WIP #34: Move AWS downloader to independent file
